tryGetSoleValue
